# First attempt at speeding up tests [WB-5097]