--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -134,6 +134,7 @@ workflows:
 
 jobs:
     test:
+        resource_class: large
         parameters:
             image:
               type: string
@@ -160,6 +161,7 @@ jobs:
             - save-tox-cache
             - save-test-results
     win:
+        resource_class: large
         parameters:
             toxenv:
               type: string
@@ -185,6 +187,7 @@ jobs:
             - save-tox-cache
             - save-test-results
     mac:
+        resource_class: large
         parameters:
             toxenv:
               type: string