--- a/setup.py
+++ b/setup.py
@@ -24,7 +24,7 @@ media_requirements = ['numpy', 'moviepy', 'pillow', 'bokeh', 'soundfile', 'plotl
 
 setup(
     name='wandb',
-    version='0.10.24.dev1',
+    version='0.10.24',
     description="A CLI and library for interacting with the Weights and Biases API.",
     long_description=readme,
     long_description_content_type="text/markdown",