--- a/tests/test_public_api.py
+++ b/tests/test_public_api.py
@@ -383,7 +383,7 @@ def test_artifact_manual_use(runner, mock_server, api):
     assert True
 
 
-def test_artifact_bracket_accessor(runner, mock_server, api):
+def test_artifact_bracket_accessor(runner, live_mock_server, api):
     art = api.artifact("entity/project/dummy:v0", type="dataset")
     assert art["t"].__class__ == wandb.Table
     assert art["s"] is None