--- a/tests/test_tb_watcher.py
+++ b/tests/test_tb_watcher.py
@@ -3,7 +3,7 @@ import platform
 import pytest
 import re
 import sys
-
+import wandb
 
 PY3 = sys.version_info.major == 3 and sys.version_info.minor >= 6
 if PY3:
@@ -106,18 +106,19 @@ def test_tb_watcher_save_row_custom_chart(mocked_run, tbwatcher_util):
     ]
 
 
-def test_tb_watcher_logdir_not_exists(mocked_run, tbwatcher_util, capsys):
+def test_tb_watcher_logdir_not_exists(runner, test_settings, tbwatcher_util, capsys):
+    run = wandb.wandb_sdk.wandb_run.Run(settings=test_settings)
     # TODO: check caplog for right error text
     pytest.importorskip("tensorboard.summary.v1")
     import tensorboard.summary.v1 as tb_summary
 
-    log_dir = os.path.join(mocked_run.dir, "test_tb_dne_dir")
+    log_dir = os.path.join(run.dir, "test_tb_dne_dir")
 
     def write_fun():
         pass
 
     _ = tbwatcher_util(
-        write_function=write_fun, logdir=log_dir, save=False, root_dir=mocked_run.dir,
+        write_function=write_fun, logdir=log_dir, save=False, root_dir=run.dir,
     )
     _, err = capsys.readouterr()
     assert err == ""