--- a/tox.ini
+++ b/tox.ini
@@ -27,7 +27,7 @@ commands_pre =
 commands =
     py{35,36,37,38,39}: ipython kernel install --user --name=wandb_python
     mkdir -p test-results
-    python -m pytest --junitxml=test-results/junit.xml --cov-config=tox.ini --cov=wandb --cov=tests/ --cov-report= --no-cov-on-fail --ignore=wandb/sweeps --ignore=build/ {posargs:tests/ wandb/sweeps/}
+    python -m pytest -n=auto --durations=20 --junitxml=test-results/junit.xml --cov-config=tox.ini --cov=wandb --cov=tests/ --cov-report= --no-cov-on-fail --ignore=wandb/sweeps --ignore=build/ {posargs:tests/ wandb/sweeps/}
 
 [testenv:py27]
 install_command = pip install -f https://download.pytorch.org/whl/torch_stable.html {opts} {packages}