--- a/wandb/__init__.py
+++ b/wandb/__init__.py
@@ -19,7 +19,7 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
-__version__ = "0.10.24"
+__version__ = "0.10.25.dev1"
 
 # Used with pypi checks and other messages related to pip
 _wandb_module = "wandb"