https://wandb.atlassian.net/browse/WB-5097

Description
-----------

Makes tests 4x+ faster than they currently are using pytest-xdist

**Before:**

<img src="https://user-images.githubusercontent.com/17/113345796-f551e480-9300-11eb-9eba-b55e3490c437.png" height="500" />

**After:**

<img src="https://user-images.githubusercontent.com/17/113345599-aad06800-9300-11eb-8327-26f540e095f0.png" height="500" />

This also exposes a number of painfully slow tests in the CI logs:

```
34.88s call     tests/test_notebooks.py::test_code_saving
24.49s call     tests/wandb_integration_test.py::test_dir_on_init_env
24.10s call     tests/test_tb_watcher.py::test_tb_watcher_logdir_not_exists
23.27s call     tests/test_tb_watcher.py::test_tb_watcher_save_row_custom_chart
22.44s call     tests/wandb_integration_test.py::test_dir_on_init_dir
20.56s call     tests/wandb_integration_test.py::test_include_exclude_config_keys
17.18s call     tests/wandb_tensorflow_test.py::test_add_pr_curve_plugin
16.38s call     tests/wandb_tensorflow_test.py::test_add_pr_curve
15.37s call     tests/wandb_tensorflow_test.py::test_compat_tensorboard
15.36s call     tests/wandb_tensorflow_test.py::test_tensorboard_log_with_wandb_log
```